# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel App API",
        "description": "Documentación de la API Travel App",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/lugares": {
            "get": {
                "tags": [
                    "Lugares"
                ],
                "summary": "Listar todos los lugares",
                "description": "Obtiene la lista de todos los lugares disponibles. Accesible tanto para usuarios registrados como no registrados",
                "operationId": "bbdab076cad171758916df3ae1190c5b",
                "responses": {
                    "200": {
                        "description": "Lista de lugares obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Lugar"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Lugares"
                ],
                "summary": "Crear un nuevo lugar",
                "description": "Solo accesible para administradores y superadministradores. Permite crear y añadir un nuevo lugar a la Base de Datos.",
                "operationId": "5589e9da706ddca3bb68cd209d44e6c9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LugarDB"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lugar creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Lugar"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acceso denegado. Solo administradores pueden crear lugares.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Solo los administradores pueden guardar un nuevo lugar"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/lugares/{id}": {
            "get": {
                "tags": [
                    "Lugares"
                ],
                "summary": "Mostrar un lugar específico (por id)",
                "description": "Obtiene los detalles de un lugar por su ID. Accesible tanto para usuarios registrados como no registrados",
                "operationId": "41225175e336077bdf12c9f169966652",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del lugar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lugar encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Lugar"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lugar no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lugar no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Lugares"
                ],
                "summary": "Actualizar un lugar",
                "description": "Permite modificar los datos de un lugar. Solo accesible para administradores y superadministradores.",
                "operationId": "a8f1a533c5c338c4da3736967d97f1b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del lugar a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LugarDB"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lugar actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Lugar"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Permiso denegado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Solo el administrador puede editar un lugar"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lugar no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lugar no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ya existe un lugar con ese nombre en nuestra BBDD."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lugares"
                ],
                "summary": "Eliminar un lugar",
                "description": "Permite eliminar un lugar. Solo accesible para administradores y superadministradores.",
                "operationId": "8c57d742ad28f86eb43a24f8314a2ce3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del lugar a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lugar borrado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lugar borrado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Permiso denegado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Solo el administrador puede borrar un lugar"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lugar no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lugar no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Lugar": {
                "title": "Lugar (Recurso API)",
                "description": "Respuesta de la API para un lugar turístico",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Cataratas del Iguazú"
                    },
                    "pais": {
                        "type": "string",
                        "example": "Argentina"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Impresionante sistema de cascadas en la frontera de Argentina y Brasil.",
                        "nullable": true
                    },
                    "imagen_url": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/iguazu.jpg",
                        "nullable": true
                    },
                    "web_url": {
                        "type": "string",
                        "format": "url",
                        "example": "https://turismo.argentina.gob.ar/iguazu",
                        "nullable": true
                    },
                    "tipoLugares": {
                        "description": "Tipos de lugar (ej: playa, montaña, parque natural...)",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "Parque natural"
                        }
                    }
                },
                "type": "object"
            },
...
            "LugarDB": {
                "title": "Lugar (BBDD)",
                "description": "Esquema de entrada/salida basado en la tabla 'lugares'",
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "Cataratas del Iguazú"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Cascadas en la selva tropical de Misiones.",
                        "nullable": true
                    },
                    "imagen_url": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/cataratas.jpg",
                        "nullable": true
                    },
                    "web_url": {
                        "type": "string",
                        "format": "url",
                        "example": "https://turismo.argentina.gob.ar/iguazu",
                        "nullable": true
                    },
                    "localizacion_url": {
                        "type": "string",
                        "format": "url",
                        "example": "https://goo.gl/maps/iguazu",
                        "nullable": true
                    },
                    "pais_id": {
                        "description": "ID del país al que pertenece el lugar",
                        "type": "integer",
                        "example": 1
                    },
                    "tipoLugares": {
                        "description": "IDs de los tipos de lugar relacionados",
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}